Add lots of prints for debugging. Basic stabilization based on last row of DCM seems to work. Need to test. Need to actually learn how DCM works in its entirety.

+++ src/flightcode/pilot.h
@@ -7,7 +7,7 @@ class Pilot {
     char commStr[];
     int serRead;
     int serInput[PACKETSIZE];
-    int axisVal[PACKETSIZE];
+    float axisVal[PACKETSIZE];
     int mapLower, mapUpper;
 //  double dir;   // Direction in radians
     bool okayToFly;
